feat: track activities api

+++ internal/services/memberships/login.go
@@ -28,7 +28,7 @@ func (s *service) Login(request memberships.LoginRequest) (string, error) {
 		return "", errors.New("invalid credentials")
 	}
 
-	accessToken, err := jwt.CreateToken(int64(userDetail.ID), userDetail.Username, s.cfg.Service.SecretJWT)
+	accessToken, err := jwt.CreateToken(userDetail.ID, userDetail.Username, s.cfg.Service.SecretJWT)
 
 	if err != nil {
 		log.Error().Err(err).Msg("fail to create JWT token")
